+++ zarrs_http/src/lib.rs
@@ -201,7 +201,7 @@ mod tests {
     use std::error::Error;
 
     const HTTP_TEST_PATH_REF: &str =
-        "https://raw.githubusercontent.com/LDeakin/zarrs/main/zarrs/tests/data/store";
+        "https://raw.githubusercontent.com/zarrs/zarrs/main/zarrs/tests/data/store";
 
     #[test]
     #[cfg_attr(miri, ignore)]